% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/net.utils.R
\title{Plant parameters of Simple Crop model}
\name{param.plant}
\alias{param.plant}

\usage{
param.plant(plant.Lfmax, plant.EMP2, plant.EMP1, plant.PD, plant.nb,plant.rm, plant.fc, plant.tb, plant.intot,plant.n,
			plant.lai, plant.w, plant.wr, plant.wc, plant.p1, plant.f1)
}

\arguments{

\item{plant.Lfmax}{Maximum number of leaves.}
\item{plant.EMP2}{Empirical coef. for expoilinear eq.}
\item{plant.EMP1}{Empirical coef. for expoilinear eq.}
\item{plant.PD}{Plant density m-2.}
\item{plant.nb}{Empirical coef. for expoilinear eq.}
\item{plant.rm}{Maximum rate of leaf appearearance (day-1).}
\item{plant.fc}{Fraction of total crop growth partitioned to canopy.}
\item{plant.tb}{Base temperature above which reproductive growth occurs (c).}
\item{plant.intot}{Duration of reproductive stage (degree days).}
\item{plant.n}{Leaf number.}
\item{plant.lai}{Canopy leaf area index (m2 m-2).}
\item{plant.w}{Total plant dry matter weight (g m-2).}
\item{plant.wr}{Root dry matter weight (g m-2).}
\item{plant.wc}{Canopy dry matter weight (g m-2).}
\item{plant.p1}{dry matter of leaves removed per plant per unit development after maximum number of leaves is reached (g).}
\item{plant.f1}{Specific leaf area (m2 g-1).}


}

\examples{
plant <- param.plant(plant.Lfmax = 12.0,plant.EMP2 = 0.64,plant.EMP1 = 0.104,plant.PD = 5.0,plant.nb = 5.3,
                     plant.rm = 0.100,plant.fc = 0.85, plant.tb = 10.0, plant.intot = 300.0,
                     plant.n = 2.0, plant.lai = 0.013, plant.w = 0.3, plant.wr = 0.045,
                     plant.wc = 0.255, plant.p1 = 0.03, plant.f1 = 0.028))
}

\keyword{plant}
