% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/net.utils.R
\title{Simulation control of Simple Crop model}
\name{param.simCtrl}
\alias{param.simCtrl}

\usage{
simCtrl(date.doyp,date.frop)
}

\arguments{

\item{date.doyp}{Date of planting (Julian day).}
\item{date.frop}{Frequency of printout (days).}

}

\examples{
	ctrl <- param.simCtrl(date.doyp = 121,date.frop = 3)
}

\keyword{simCtrl}
