% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/net.utils.R
\title{Soil parameters of Simple Crop model}
\name{param.soil}
\alias{param.soil}

\usage{
param.soil(soil.WPp, soil.FCp, soil.STp, soil.DP,soil.DRNp, soil.CN, soil.SWC)
}

\arguments{

\item{soil.WPp}{Water content at wilting point (fraction of void space).}
\item{soil.FCp}{Water content at field capacity (fraction of void space).}
\item{soil.STp}{Water content saturation (fraction of void space).}
\item{soil.DP}{Depth of the profile considered in the simulation (cm).}
\item{soil.DRNp}{Daily drainage percentage (fraction of void space).}
\item{soil.CN}{Runoff curve number.}
\item{soil.SWC}{Actual soil water storage in the profile (mm).}


}

\examples{
soil <- param.soil(soil.WPp = 0.06,soil.FCp = 0.17,soil.STp = 0.28, soil.DP = 145.000,
                   soil.DRNp = 0.10, soil.CN = 55.00,soil.SWC = 246.5)
}

\keyword{soil}
