% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/net.utils.R
\title{Execute Simple Crop model}
\name{runSimpleCrop}
\alias{runSimpleCrop}

\usage{
runSimpleCrop(weather, plant, soil, irrig, simCtrl)
}

\arguments{

\item{weather}{receives a data frame that contains the day of the year along with the maximum and minimum temperature, rainfall, solar radiation, PAR.}
\item{plant}{See \code{\link{param.plant}}.}
\item{soil}{See \code{\link{param.soil}}.}
\item{irrig}{Receives a data frame containing the day of the year along with the irrigation amount.}
\item{simCtrl}{See \code{\link{param.simCtrl}}.}

}

\details{The execution of model returns a list with three data frames, water balance, plant, soil water
}

\examples{
out<-runSimpleCrop(weather = weather,plant = plant,soil = soil,irri = irrig,simCtrl = ctrl)
wbal <- out$wbal
sw <- out$sw
plant <-out$plant
}

\keyword{runSimpleCrop}
